llm-invoker-service/requirements.txt
openai>=2.6.1
anthropic>=0.72.0
fastapi>=0.120.4
pydantic>=2.12.3
uvicorn>=0.38.0
python-dotenv>=1.2.1